backend/pyproject.toml
[tool.poetry]
name = "meeting-transcription-assistant"
version = "0.1.0"
description = ""
authors = ["rainsound.ai <hello@rainsound.ai>"]
readme = "README.md"
package-mode = false
[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.1"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0"
pre-commit = "^3.7.0"
whisper-openai = "^1.0.0"
pydub = "^0.25.1"
python-multipart = "^0.0.11"
openai = "^1.50.2"

[tool.poetry.dependencies.uvicorn]
extras = ["standard"]
version = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
